refactor(types): rename OrderRecord to OrderSchema and remove list/dashboard types
