+++ packages/tv-ad-admin-next/components/list/order-group.tsx
@@ -1,6 +1,6 @@
 import { OrderRow } from './order-row'
 
-import { type OrderRecordForList } from '@/types/order'
+import { type OrderRecordForList } from '@/graphql/queries/orders'
 
 type OrderGroupProps = {
   orders: OrderRecordForList[]
