+++ packages/tv-ad-admin-next/types/order.ts
@@ -1,36 +1,29 @@
 import { type OrderState } from '@/constants'
 
-export type OrderRecord = {
+export type OrderSchema = {
   id: number
   orderNumber: string
   name: string | null
+  nameEditable?: boolean
+  member?: { id: string } | null
   scheduleStartDate: Date | null
   scheduleEndDate: Date | null
   scheduleStartDateString?: string | null
   scheduleEndDateString?: string | null
+  // Keystone `schedule` field (free-text). Frontend may also derive start/end above.
+  schedule?: string | null
+  scheduleEditable?: boolean
   state: OrderState
-  relatedOrder?: OrderRecord
+  // Keystone `relatedOrder` is many=true
+  relatedOrder?: { id: string }[]
+  attachment?: { id: string } | null
+  paragraphOne?: string | null
+  paragraphOneEditable?: boolean
+  paragraphTwo?: string | null
+  paragraphTwoEditable?: boolean
+  image?: { id: string } | null
+  imageEditable?: boolean
+  demoImage?: { id: string }[]
   createdAt: string
   updatedAt: string
 }
-
-export type OrderRecordForList = Pick<
-  OrderRecord,
-  | 'id'
-  | 'orderNumber'
-  | 'name'
-  | 'state'
-  | 'scheduleStartDate'
-  | 'scheduleEndDate'
-  | 'scheduleStartDateString'
-  | 'scheduleEndDateString'
-  | 'createdAt'
-  | 'updatedAt'
-> & {
-  relatedOrder?: { id: string }
-}
-
-export type OrderRecordForDashboard = Pick<
-  OrderRecord,
-  'id' | 'state' | 'updatedAt'
->
